+++ app/src/main/java/com/thondar/utils/Test.java
@@ -1,0 +1,9 @@
+package com.thondar.utils;
+
+import io.github.ouyuanx.androidutils.permission.PermissionUtils;
+
+final class Test {
+
+    void test() {
+    }
+}